fix cost tracking by tags
